refactor: clean up codebase with various improvements
- Remove unused imports and dead code annotations
- Simplify mathematical operations and loops
- Implement Default trait for multiple structs
- Fix clippy warnings and improve code style
- Consolidate configuration initialization
- Optimize memory usage and performance
- Standardize error handling patterns
- Improve documentation and type safety

+++ examples/swe_3d_liver_fibrosis.rs
@@ -166,7 +166,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         solver.propagate_volumetric_waves_with_body_forces(&body_forces, &push_times, &sources)?;
     println!(
         "   Simulated {} time steps of wave propagation",
-        displacement_history.len() as usize
+        displacement_history.len()
     );
     println!("   Wave front tracking completed\n");
